DB와 연결되어있는 서비스 부분 수정

+++ app/src/main/java/comkimhyeockjin/github/termproject/LocationInfo.java
@@ -13,9 +13,9 @@ public class LocationInfo {
     public String date;
     public double lng;
     public double lat;
-    public double time;
+    public int time;
 
-    public LocationInfo(int id, String date, double lng, double lat, double time) {
+    public LocationInfo(int id, String date, double lng, double lat, int time) {
         this.id = id;
         this.date = date;
         this.lng = lng;
@@ -55,7 +55,7 @@ public class LocationInfo {
         this.lat = lat;
     }
 
-    public double getTime() {
+    public int getTime() {
         return time;
     }
